requirements.txt
hikari==2.0.0.dev122
python-dotenv==1.0.1
tinydb==4.8.0
